Move ActuatorTests to relevant Package and refactor LdapHealthServiceTest
